+++ dap-etl/extraction/src/main/scala/org/broadinstitute/monster/dap/HLESurveyExtractionPipeline.scala
@@ -6,7 +6,8 @@ import org.broadinstitute.monster.dap.common.{
   FilterDirective,
   FilterOps,
   HttpWrapper,
-  OkWrapper
+  OkWrapper,
+  RedCapClient
 }
 
 import java.time.{OffsetDateTime, ZoneOffset}
