+++ dap-etl/extraction/src/main/scala/org/broadinstitute/monster/dap/common/ExtractionPipelineBuilder.scala
@@ -7,7 +7,6 @@ import org.apache.beam.sdk.coders.{KvCoder, StringUtf8Coder}
 import org.apache.beam.sdk.transforms.{GroupIntoBatches, ParDo}
 import org.apache.beam.sdk.values.KV
 import org.broadinstitute.monster.common.{PipelineBuilder, StorageIO}
-import org.broadinstitute.monster.dap._
 import org.slf4j.LoggerFactory
 import upack.Msg
