+++ dap-etl/extraction/src/test/scala/org/broadinstitute/monster/dap/CslbExtractionPipelineBuilderIntegrationSpec.scala
@@ -2,7 +2,7 @@ package org.broadinstitute.monster.dap
 
 import java.time.{LocalDate, LocalTime, OffsetDateTime, ZoneOffset}
 import better.files.File
-import common.Args
+import common.{Args, GetRecords, RedcapRequestGeneratorParams}
 import org.broadinstitute.monster.common.PipelineBuilderSpec
 
 class CslbExtractionPipelineBuilderIntegrationSpec extends PipelineBuilderSpec[Args] {
@@ -36,7 +36,7 @@ class CslbExtractionPipelineBuilderIntegrationSpec extends PipelineBuilderSpec[A
     )
   )
 
-  val followUpRecords = RedcapRequestGeneratorParams(
+  val followUpRecords = common.RedcapRequestGeneratorParams(
     testArgs.apiToken,
     CslbExtractionPipeline.arm,
     GetRecords(
