+++ dap-etl/extraction/src/test/scala/org/broadinstitute/monster/dap/RedCapClientSpec.scala
@@ -3,7 +3,13 @@ package org.broadinstitute.monster.dap
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
 import okhttp3.FormBody
-import org.broadinstitute.monster.dap.common.{FilterDirective, FilterOps}
+import org.broadinstitute.monster.dap.common.{
+  FilterDirective,
+  FilterOps,
+  GetRecords,
+  RedCapClient,
+  RedcapRequestGeneratorParams
+}
 
 class RedCapClientSpec extends AnyFlatSpec with Matchers {
